added test date and exception in the controllers

+++ src/main/java/com/example/transport/service/JourneyStopServiceImpl.java
@@ -2,7 +2,7 @@ package com.example.transport.service;
 
 
 import com.example.transport.domain.JourneyStop;
-import com.example.transport.exception.TransportException;
+import com.example.transport.exception.JourneyStopException;
 import com.example.transport.repository.JourneyStopRepo;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -17,20 +17,25 @@ public class JourneyStopServiceImpl {
     public List<JourneyStop> getAllJourney(){
         return journeyStopRepo.findAll();
     }
+
     public JourneyStop getJourneyById(int id){
-        return journeyStopRepo.findById(id).orElseThrow(()-> new TransportException("JourneyStop with id " + id + " was not found"));
+        return journeyStopRepo.findById(id).orElseThrow(()-> new JourneyStopException("JourneyStop with id " + id + " was not found"));
     }
+
     public void addNewJourneyStop(JourneyStop journeyStop){
         journeyStopRepo.save(journeyStop);
     }
+
     public void updateJourneyStop(int id, JourneyStop journeyStop){
         JourneyStop oldJourneyStop = journeyStopRepo.getReferenceById(id);
-       // oldJourneyStop.setId(journeyStop.getId());
+        oldJourneyStop.setId(journeyStop.getId());
         journeyStopRepo.save(oldJourneyStop);
     }
+
     public void deleteJourneyStop(int id){
+        if (!journeyStopRepo.existsById(id)){
+            throw new JourneyStopException("JourneyStop with id " + id + " not found");
+        }
         journeyStopRepo.deleteById(id);
     }
-
-
 }
