+++ src/main/java/com/example/transport/service/StopTransportServiceImpl.java
@@ -1,7 +1,7 @@
 package com.example.transport.service;
 
 import com.example.transport.domain.StopTransport;
-import com.example.transport.exception.TransportException;
+import com.example.transport.exception.StopTransportNotFound;
 import com.example.transport.repository.StopTransportRepo;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -16,18 +16,26 @@ public class StopTransportServiceImpl {
     public List<StopTransport> getAllStopTransport(){
         return stopTransportRepo.findAll();
     }
+
     public void addNewStopTransport(StopTransport stopTransport){
         stopTransportRepo.save(stopTransport);
     }
-    public void updateStopTransport(int id,StopTransport stopTransport) {
+
+    public void updateStopTransport(String id,StopTransport stopTransport) {
         StopTransport oldStopTransport = stopTransportRepo.getReferenceById(id);
         oldStopTransport.setId(stopTransport.getId());
         stopTransportRepo.save(oldStopTransport);
     }
-    public void deleteStopTransport(int id){
+
+    public void deleteStopTransport(String id){
+        if(!stopTransportRepo.existsById(id)) {
+            throw new StopTransportNotFound("Stop transport with id " + id + " not found");
+
+        }
         stopTransportRepo.deleteById(id);
     }
-    public StopTransport getStopTransportById(int id){
-        return   stopTransportRepo.findById(id).orElseThrow(()-> new TransportException("StopTransport with numberStop " + id + " was not found"));
+
+    public StopTransport getStopTransportById(String id){
+        return   stopTransportRepo.findById(id).orElseThrow(()-> new StopTransportNotFound("StopTransport with numberStop " + id + " was not found"));
     }
 }
